Migrated to gson from moshi

+++ app/src/main/java/com/bondidos/wotstatisticbybondidos/data/entiyes/WotUser.kt
@@ -1,4 +1,5 @@
 package com.bondidos.wotstatisticbybondidos.data.entiyes
+/*
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
@@ -14,6 +15,7 @@ data class WotUser (
     @Json(name = "nickname")
     val nickname : String,
     @Json(name = "account_id")
-    val account_id : Long
+    val account_id : Int
 )
+*/
